api-functional test, dependencies fixes
