+++ InventoryDefaultForCountrySourceSelection/composer.json
@@ -4,9 +4,9 @@
     "require": {
         "php": "~7.3.0||~7.4.0",
         "magento/framework": "*",
-        "magento/module-inventory-distance-based-source-selection-api": "1.1.*",
-        "magento/module-inventory-source-selection-api": "1.3.*",
-        "magento/module-inventory-api": "1.1.*",
+        "magento/module-inventory-distance-based-source-selection-api": "*",
+        "magento/module-inventory-source-selection-api": "*",
+        "magento/module-inventory-api": "*",
         "magento/module-config": "*"
     },
     "type": "magento2-module",
@@ -22,5 +22,5 @@
             "Magento\\InventoryDefaultForCountrySourceSelection\\": ""
         }
     },
-    "version": "1.1.0"
+    "version": "1.2.0"
 }
