+++ InventoryDefaultForCountrySourceSelectionAdminUi/composer.json
@@ -3,7 +3,7 @@
     "description": "N/A",
     "require": {
         "php": "~7.3.0||~7.4.0",
-        "magento/framework": "*"
+        "magento/module-inventory-default-for-country-source-selection": "*"
     },
     "type": "magento2-module",
     "license": [
@@ -18,5 +18,5 @@
             "Magento\\InventoryDefaultForCountrySourceSelectionAdminUi\\": ""
         }
     },
-    "version": "1.1.0"
+    "version": "1.2.0"
 }
