+++ InventoryDefaultForCountrySourceSelectionAdminUi/etc/module.xml
@@ -8,5 +8,8 @@
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
         xsi:noNamespaceSchemaLocation="urn:magento:framework:Module/etc/module.xsd">
     <module name="Magento_InventoryDefaultForCountrySourceSelectionAdminUi">
+        <sequence>
+            <module name="Magento_InventoryDefaultForCountrySourceSelection" />
+        </sequence>
     </module>
 </config>
